module completeGradleProperties is renamed to completion, some classes are renamed
