+++ org.eclipse.buildship.kts.scripts/language-server-for-gradle-properties/app/src/main/java/server/completion/PropertyModelJson.java
@@ -1,4 +1,4 @@
-package server.completeGradleProperties;
+package server.completion;
 
 import java.util.ArrayList;
